+++ mobile/android/android-components/docs/changelog.md
@@ -15,6 +15,10 @@ permalink: /changelog/
 * **browser-errorpages**
   * Added strings for "no network connection" error pages
 
+* **browser-menu**
+  * Replaced `BrowserMenuHighlightableItem.Highlight` with `BrowserMenuHighlight.HighPriority` to highlight a menu item with some background color. `Highlight` has been deprecated.
+  * Added `BrowserMenuHighlight.LowPriority` to highlight a menu item with a dot over the icon.
+
 # 23.0.0
 
 * [Commits](https://github.com/mozilla-mobile/android-components/compare/v22.0.0...v23.0.0)
@@ -37,7 +41,7 @@ permalink: /changelog/
   * Account profile cache is now used, removing a network call from most instances of account manager instantiation.
   * Fixed a bug where account would disappear after restarting an app which hit authentication problems.
   * Deprecated the `StorageSync` class. Please use the `SyncManager` class instead.
-  
+
 * **service-glean**
   * Glean was updated to v21.2.0
     * Two new metrics were added to investigate sending of metrics and baseline pings.